requirements.txt
asn1crypto==1.0.1
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.18.0
idna==2.8
pycparser==2.19
python-telegram-bot==12.1.1
requests==2.22.0
six==1.12.0
tornado==6.0.3
urllib3==1.25.6
virtualenv==16.7.5
vk-api==11.5.1
apiai==1.2.3
